requirements.txt
azure-cli
streamlit
streamlit-authenticator
streamlit-extras
st-pages
openai
pandas
openpyxl
langchain
tiktoken
PyPDF2
azure-identity
azure-search-documents==11.4.0
chromadb
pydub
azure-storage-blob  